test-requirements.txt
-c package_constraints.txt
-r requirements.txt

aiohttp==3.8.4
coverage==7.2.3
pytest==7.3.1
pytest-asyncio==0.20.3
pytest-cov==3.0.0
pytest-homeassistant-custom-component==0.13.30
pytest-xdist==3.2.1
pyyaml==6.0
deepdiff
mutagen==1.46.0
home-assistant-intents==2023.4.26
webrtcvad==2.0.10

# for sonos
soco==0.29.1
async_upnp_client==0.33.1
spotipy==2.23.0
plexapi==4.13.2
plexwebsocket==0.0.13
# sonos_websocket
hassil==1.0.6
sonos-websocket==0.1.1

# for mobile app
PyNaCl==1.5.0
